requirements.txt
appdirs==1.4.3
apturl==0.5.2
argon2-cffi==19.2.0
asgiref==3.2.3
asn1crypto==0.24.0
bcrypt==3.1.6
beautifulsoup4==4.8.2
blinker==1.4
Brlapi==0.6.7
bs4==0.0.1
certifi==2018.8.24
cffi==1.14.0
chardet==3.0.4
Click==7.0
command-not-found==0.3
configobj==5.0.6
cryptography==2.6.1
cupshelpers==1.0
dbus-python==1.2.12
decorator==4.4.1
defer==1.0.6
distlib==0.3.0
distro==1.3.0
distro-info===0.21ubuntu4
dj-database-url==0.5.0
Django==3.0.4
django-appconf==1.0.3
django-countries==5.5
django-crispy-forms==1.8.1
django-geoip==0.5.2.1
django-heroku==0.3.1
django-mapbox-location-field==1.4.0
django-user-activity-log2==0.0.24
djangorestframework==3.11.0
duplicity==0.8.4
entrypoints==0.3
fasteners==0.12.0
filelock==3.0.12
future==0.16.0
geoip2==3.0.0
gunicorn==20.0.4
html5lib==1.0.1
httplib2==0.11.3
idna==2.6
importlib-metadata==1.5.0
Js2Py==0.67
keyring==18.0.1
keyrings.alt==3.1.1
language-selector==0.1
launchpadlib==1.10.7
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.10.0
lxml==4.5.0
macaroonbakery==1.2.3
Mako==1.0.7
MarkupSafe==1.1.0
maxminddb==1.5.2
monotonic==1.5
netifaces==0.10.4
oauthlib==2.1.0
olefile==0.46
paramiko==2.6.0
pexpect==4.6.0
Pillow==6.1.0
protobuf==3.6.1
psutil==5.5.1
psycopg2==2.8.4
psycopg2-binary==2.8.4
py-mini-racer==0.1.18
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
PyGObject==3.34.0
pyjsparser==2.7.1
PyJWT==1.7.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==1.9.0+ubuntu1.3
python-dateutil==2.7.3
python-debian==0.1.36
pytz==2019.2
pyxdg==0.25
PyYAML==5.1.2
ratelim==0.1.6
reportlab==3.5.23
requests==2.22.0
requests-unixsocket==0.1.5
scapy==2.4.3
SecretStorage==2.3.1
simplejson==3.16.0
six==1.12.0
soupsieve==2.0
sqlparse==0.3.0
system-service==0.3
systemd-python==234
termcolor==1.1.0
tzlocal==2.0.0
ubuntu-advantage-tools==19.5
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
vboxapi==1.0
virtualenv==20.0.10
wadllib==1.3.3
webencodings==0.5.1
whitenoise==5.0.1
xkit==0.0.0
zipp==3.1.0
